Sorted List
A linked list implementation in order to
create a sorted list that accepts void pointers (for information hiding)
and then
casts the passed arguments into the corresponding type. We used an
iterator to
traverse through the linked list and used the comparator function to
determine
where the node is inserted. The program outputs the linked list in
sorted form in
descending order of numbers and characters depending on the specific
input.